Add limit graph and equation 4
Added a graph to numerically clarify the 'zeros' of the distance function derived in example 4.
Those who wish to replicate the graph, should comment out the current function and uncomment
the one above. These two functions were made for simplicity sake, as fsolve searches for the
zeros of a function and graphing it, we wish to keep lhs and rhs terms as intact as possible.